backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
flasgger==0.9.7.1
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1